Apply the wasm theme lookup fix to theme package too
