+++ theme/theme_wasm.go
@@ -1,0 +1,9 @@
+//go:build wasm || test_web_driver
+
+package theme
+
+import "fyne.io/fyne/v2"
+
+func setupSystemTheme(fallback fyne.Theme) fyne.Theme {
+	return fallback
+}
